Change initialization, argument parsing for heroku
This previously didn't work with heroku since gunicorn doesn't run my
application with __main__, and thus doesn't initialize the twitter api,
etc.

Therefore, I make it *required* that the API keys for Twitter be
defined in the environment such that it is always initialized. I tone
down the argument parsing accordingly.

This makes it possible to run the application from both heroku/foreman
with gunicorn and the command-line with `python TwitterSA.py` (provided
that environment variables are set correctly)