Add incomplete sentiment, classifiers boilerplate
Far, far from complete, just saving work -

sentiment.py is the interface between the actual hard-coded data
processing of classification.py and the Flask web application, defining
several classes to facilitate the visualization and handling of data.

classifiers.py includes a lot of the data preprocessing required for
data loading and classification, including a tweet preprocess function
and the basic counting of a Naive Bayes Classifier assuming binary
features.

Later I'll be switching to a more general, normal Bayes classifier and
dealing with sparse data with laplace smoothing or something similar.

NLTK is a good model for this.

see the TODO FIXME OPTIMIZE...etc for next steps.

sentiment import added to TwitterSA